pyproject.toml
[project]
name = "jurybot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.7.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
]
